backend/requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
uvicorn[standard]==0.24.0